Add rake task to test if db is accepting conns
The script to run the tests on Jenkins starts the database container,
and then starts running the migrations from the app container. The
problem occurs when most of the app container is cached by docker and
starts before the database has come up.

This rake task can be called before any migrations and will essentially
just return an exit code of 0 if it connects to the database. It will
retry and ultimately exit with a non 0 code if it can't connect allowing
the Jenkins job to fail.